requirements.txt
numpy==2.2.4
sounddevice
python-telegram-bot
spidev
keybow
